Gave up.
In the lambda case for lifting, we need to know the kind of the
expression in the lambda. The only way to do this is to thread through
the outer typing context so we can do lookup. Besides, once I write
down the lifting algorithm, I still need to write the proof.

It's just not worth it. I'm confident I could, but I won't. Not
today.